Add support to ppc64le for python 3.7 and 3.8
Adds CI support for the IBM Power Little Endian (ppc64le) architecture.
The idea is to ensure that the builds on this architecture are
continuously tested along with the Intel builds (amd64) as this is part
of the ubuntu distro on that architecture as well and detecting (and
fixing) any issues or failures early would help to ensure that we are
always up to date. Install libmagickwand-dev for ppc64le